style: move columns.ts out of general purpose Table folder
